movies_admin/requirements/dev.txt
-r base.txt
django-debug-toolbar==2.2  # Конфигурируемая панель с информацией для отладки проекта
